Fix CPU spike in event server
If the client socket disconnects in a particular way, the event server can get
stuck in the inner while loop where it reads protobuf messages in. The existing
error check for a null protobuf would break that iteration of the loop, but did
not break out of the loop itself.

Changed 'continue' to 'break' as a temporary fix. It really should throw an
exception, but this is good enough for now since I'm planning to rewrite the
event server pretty soon anyway.

Change-Id: Ic4f5a32beff2688a38bb373a882f70a477c4a75b