Workaround to stop video stream on client disconnect
The merging of fbstream and the event server left us with no way to
stop sending video when a client disconnects. Previously, the event
server would send a BYE message to the separate fbstream, after which
fbstream would exit and restart. Since we can't do that anymore and I
still haven't figure out the right way to tell webrtc to stop, this
patch has the event server itself exit and get restart by Android.

Change-Id: Iebb1fbdf0cb4645a8ca43a157b194188894e509d