Fixed KitKat svmpd boot issue
The receiver was not receiving the BOOT_COMPLETED intent broadcast,
and the svmpd service was not starting. This was happening because
KitKat seems to handle initial app installs differently than ICS.
As such, on first boot there was no default Launcher preference set
and that prevented some packages from receiving intent broadcasts.
Updated the API to 19, updated makefile for KitKat-related changes,
disabled the svmpd LauncherActivity in the manifest, and added code
to programmatically enable that activity after boot completed.