Another svmpd boot issue fix
Ever since upgrading to the KitKat ROM, it appears that the AOSP
launcher is not selected as the "default" launcher. When multiple
launchers were installed, this somehow prevented the BOOT_COMPLETED
broadcast from being sent properly. To work around this, we
removed the SVMP Single App Mode launcher from the Manifest and
instead enabled the component programmatically when the broadcast
was received. This worked, however, rebooting the VM without first
choosing a default launcher (e.g. logging in with a client) would
cause the same BOOT_COMPLETED broadcast issue. Now, after we enable
the SVMP launcher, we also set the AOSP launcher to default. This
fixes the recurring boot issue.